requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
motor==3.3.1
aiohttp>=3.9.0
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
requests>=2.31.0
python-multipart>=0.0.9
